Add files via upload
Rethought the way we need to extract data from the json files. I think we need to take normalization into consideration when dealing with the following keys
- tech_self_score
- strengths
- weaknesses

I think we need to create seperate tables/dataframes for these values. I've made some headway with this in that I've created a function that would be able to create the headers needed for each table, but the next step is actually inserting the values into these specific dataframes